+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "Background Phi Colors" extension will be documented 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 0.?.? - 2019-03-2?
+
+### Added
+
+- `background-phi-colors.enabledPanels` setting
+
 ## 0.2.0 - 2019-03-20
 
 ### Added